+++ Sources/SwiftNPSDataModels/Module.swift
@@ -1,6 +1,0 @@
-// Response models and endpoint values will live in this transport-independent module.
-#if canImport(FoundationEssentials)
-import FoundationEssentials
-#else
-import Foundation
-#endif